[GraphBolt] split OnDiskDataset into separate file
