+++ python/dgl/graphbolt/__init__.py
@@ -12,6 +12,7 @@ from .feature_store import *
 from .feature_fetcher import *
 from .copy_to import *
 from .dataset import *
+from .ondisk_dataset import *
 from .dataloader import *
 from .subgraph_sampler import *
